requirements.txt
requests
django-fullurl